Simplify protocol
Previously `Setup`, `SetupResponse`, and `Error` all accepted only a
single body frame, rather than multiple frames. We now simplify the
protocol to make all `body` passing optionally framed. If `body` is a
list/tuple, we deserialize it as a list of frames (as before). If `body`
is a buffer, we deserialize it as a buffer, avoiding the creation of the
list. This simplifies the implementation, and allows for more uniform
treatment of the `body` kwarg in user-facing methods.